Add optimization stage

+++ source/lib/vm.cpp
@@ -172,7 +172,7 @@ void thread_t::do_INS_RET_() {
   // pop return value
   const int32_t sval = pop();
   // remove arguments and local vars
-  if (s_head_ < operand) {
+  if (int32_t(s_head_) < operand) {
     set_error_(thread_error_t::e_stack_underflow);
   }
   else {
